IA-3319 Refactoring
