+++ plugins/polio/js/src/domains/VaccineModule/StockManagement/StockVariation/Modals/CreateEditFormA.tsx
@@ -16,7 +16,7 @@ import { DateInput, NumberInput } from '../../../../../components/Inputs';
 import { useCampaignOptions, useSaveFormA } from '../../hooks/api';
 import { EditIconButton } from '../../../../../../../../../hat/assets/js/apps/Iaso/components/Buttons/EditIconButton';
 import { useFormAValidation } from './validation';
-import { accept, processErrorDocsBase } from '../../../SupplyChain/Details/VaccineRequestForm/VaccineRequestForm';
+import { acceptPDF, processErrorDocsBase } from '../../../SupplyChain/Details/utils';
 
 type Props = {
     formA?: any;
@@ -124,7 +124,7 @@ export const CreateEditFormA: FunctionComponent<Props> = ({
                 </Box>
                 <Box mb={2}>
                     <FilesUpload
-                        accept={accept}
+                        accept={acceptPDF}
                         files={formik.values.document ? [formik.values.document] : []}
                         onFilesSelect={files => {
                             if (files.length) {
