+++ plugins/polio/js/src/domains/VaccineModule/SupplyChain/Details/utils.ts
@@ -8,6 +8,7 @@ import {
     UseHandleSubmitArgs,
     VAR,
 } from '../types';
+import { Accept } from 'react-dropzone';
 
 type SaveAllArgs = {
     changedTabs: TabValue[];
@@ -199,3 +200,14 @@ export const makeHandleSubmit =
             },
         );
     };
+
+
+export const acceptPDF: Accept = {
+    'application/pdf': ['.pdf'],
+};
+
+export const processErrorDocsBase = (err_docs) => {
+    if (!err_docs) return [];
+    if (!Array.isArray(err_docs)) return [err_docs]; 
+    else return err_docs;
+};
